pyproject.toml
[project]
name = "animations"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.3.5",
    "pillow>=10.0.0",
    "pygame>=2.6.1",
    "pyright>=1.1.407",
    "raylib>=5.5.0.4",
    "ruff>=0.14.9",
]
